Background section changes from Ryan
This is the only area I had questions about @ryansosin - which I sent on slack. **Step 1:** The user is taken to the Carrier Discovery UI website, where they choose the carrier associated with their primary device. If the user is authorizing a secondary device from an app on a tablet, the SDK will use a webview for this step.